rb_proc_t: reduce to 64 bytes from 72 on 64-bit
* vm_core.h (rb_proc_t): reduce to 64 bytes from 72 on 64-bit

This allows rb_proc_t to fit inside a single cache line on x86-64

git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@46886 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ vm_core.h
@@ -738,9 +738,9 @@ typedef struct {
 
     VALUE envval;		/* for GC mark */
     VALUE blockprocval;
-    int safe_level;
-    int is_from_method;
-    int is_lambda;
+    int8_t safe_level;		/* 0..4 */
+    int8_t is_from_method;	/* bool */
+    int8_t is_lambda;		/* bool */
 } rb_proc_t;
 
 #define GetEnvPtr(obj, ptr) \
